+++ src/Telegram/Types/RichMessage/RichBlock/RichBlockTableCell.php
@@ -24,7 +24,7 @@ class RichBlockTableCell extends BaseType
      * Optional. Text in the cell. If omitted, then the cell is invisible.
      * @var string|RichText[]|RichText
      */
-    #[ArrayType(RichText::class)]
+    #[ArrayType(RichText::class, skipScalars: true)]
     #[RichTextUnionResolver]
     public string|array|RichText|null $text = null;