+++ src/Telegram/Types/RichMessage/RichText/RichTextMention.php
@@ -26,7 +26,7 @@ class RichTextMention extends BaseType implements RichText
      * The text
      * @var string|RichText[]|RichText
      */
-    #[ArrayType(RichText::class)]
+    #[ArrayType(RichText::class, skipScalars: true)]
     #[RichTextUnionResolver]
     public string|array|RichText $text;